update grpc command service
